requirements.txt
numpy>=1.21.0
matplotlib>=3.5.0
opencv-python>=4.5.0
Pillow>=8.0.0
scipy>=1.7.0
scikit-image>=0.18.0
tifffile>=2021.0.0
